+++ src/Data/Map/Monoidal/Strict.hs
@@ -1,11 +1,11 @@
 {-# LANGUAGE CPP #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE DeriveTraversable #-}
 {-# LANGUAGE DeriveDataTypeable #-}
-{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
 
 -- | This module provides a 'Data.Map' variant which uses the value's
@@ -28,6 +28,7 @@ module Data.Map.Monoidal.Strict
     , assocs
     , elems
     , keys
+    , (!?)
     , (!)
     , (\\)
     , adjust
@@ -35,6 +36,11 @@ module Data.Map.Monoidal.Strict
     , alter
     , delete
     , deleteAt
+    , take
+    , drop
+    , splitAt
+    , lookupMin
+    , lookupMax
     , deleteFindMax
     , deleteFindMin
     , deleteMax
@@ -46,6 +52,8 @@ module Data.Map.Monoidal.Strict
     , empty
     , filter
     , filterWithKey
+    , restrictKeys
+    , withoutKeys
     , findIndex
     , findMax
     , findMin
@@ -63,6 +71,10 @@ module Data.Map.Monoidal.Strict
     , fromAscListWithKey
     , fromDistinctAscList
     , fromDistinctList
+    , fromDescList
+    , fromDescListWith
+    , fromDescListWithKey
+    , fromDistinctDescList
     , fromList
     , fromListWith
     , fromListWithKey
@@ -104,6 +116,9 @@ module Data.Map.Monoidal.Strict
     , null
     , partition
     , partitionWithKey
+    , takeWhileAntitone
+    , dropWhileAntitone
+    , spanAntitone
     , split
     , splitLookup
     , splitRoot
@@ -111,6 +126,7 @@ module Data.Map.Monoidal.Strict
     , toDescList
     , toList
     , traverseWithKey
+    , traverseMaybeWithKey
     , unionWith
     , unionWithKey
     , unionsWith
@@ -127,19 +143,19 @@ module Data.Map.Monoidal.Strict
     -- , showTreeWith
     ) where
 
-import Prelude hiding (null, lookup, map, foldl, foldr, filter)
+import Prelude hiding (null, lookup, map, foldl, foldr, filter, take, drop, splitAt)
 
 import Data.Coerce (coerce)
+import Data.Set (Set)
 import Data.Semigroup
 import Data.Foldable (Foldable)
 import Data.Traversable (Traversable)
 import Control.Applicative (Applicative, pure)
 import Data.Data (Data)
 import Data.Typeable (Typeable)
-import Data.Set (Set)
 
 #if MIN_VERSION_base(4,7,0)
-import qualified GHC.Exts  as IsList
+import qualified GHC.Exts as IsList
 #endif
 
 import Control.DeepSeq
@@ -293,6 +309,11 @@ keys :: MonoidalMap k a -> [k]
 keys = M.keys . unpack
 {-# INLINE keys #-}
 
+(!?) :: forall k a. Ord k => MonoidalMap k a -> k -> Maybe a
+(!?) m k = (coerce m) !? k
+infixl 9 !?
+{-# INLINE (!?) #-}
+
 
 (!) :: forall k a. Ord k => MonoidalMap k a -> k -> a
 (!) = coerce ((M.!) :: M.Map k a -> k -> a)
@@ -418,6 +439,10 @@ traverseWithKey :: Applicative t => (k -> a -> t b) -> MonoidalMap k a -> t (Mon
 traverseWithKey = itraverse
 {-# INLINE traverseWithKey #-}
 
+traverseMaybeWithKey :: forall f k a b. Applicative f => (k -> a -> f (Maybe b)) -> MonoidalMap k a -> f (MonoidalMap k b)
+traverseMaybeWithKey f m = coerce <$> M.traverseMaybeWithKey f (coerce m)
+{-# INLINE traverseMaybeWithKey #-}
+
 mapAccum :: forall k a b c. (a -> b -> (a, c)) -> a -> MonoidalMap k b -> (a, MonoidalMap k c)
 mapAccum = coerce (M.mapAccum :: (a -> b -> (a, c)) -> a -> M.Map k b -> (a, M.Map k c))
 {-# INLINE mapAccum #-}
@@ -430,8 +455,8 @@ mapAccumRWithKey :: forall k a b c. (a -> k -> b -> (a, c)) -> a -> MonoidalMap 
 mapAccumRWithKey = coerce (M.mapAccumRWithKey :: (a -> k -> b -> (a, c)) -> a -> M.Map k b -> (a, M.Map k c))
 {-# INLINE mapAccumRWithKey #-}
 
-mapKeys :: forall k1 k2 a. (Ord k2, Semigroup a) => (k1 -> k2) -> MonoidalMap k1 a -> MonoidalMap k2 a
-mapKeys = mapKeysWith (<>)
+mapKeys :: forall k1 k2 a. Ord k2 => (k1 -> k2) -> MonoidalMap k1 a -> MonoidalMap k2 a
+mapKeys = coerce (M.mapKeys :: (k1 -> k2) -> M.Map k1 a -> M.Map k2 a)
 {-# INLINE mapKeys #-}
 
 mapKeysWith :: forall k1 k2 a. Ord k2 => (a -> a -> a) -> (k1 -> k2) -> MonoidalMap k1 a -> MonoidalMap k2 a
@@ -508,8 +533,8 @@ toList :: forall k a. MonoidalMap k a -> [(k, a)]
 toList = coerce (M.toList :: M.Map k a -> [(k, a)])
 {-# INLINE toList #-}
 
-fromList :: forall k a. (Ord k, Semigroup a) => [(k, a)] -> MonoidalMap k a
-fromList = fromListWith (<>)
+fromList :: forall k a. Ord k => [(k, a)] -> MonoidalMap k a
+fromList = coerce (M.fromList :: [(k, a)] -> M.Map k a)
 {-# INLINE fromList #-}
 
 fromListWith :: forall k a. Ord k => (a -> a -> a) -> [(k, a)] -> MonoidalMap k a
@@ -528,8 +553,8 @@ toDescList :: forall k a. MonoidalMap k a -> [(k, a)]
 toDescList = coerce (M.toDescList :: M.Map k a -> [(k, a)])
 {-# INLINE toDescList #-}
 
-fromAscList :: forall k a. (Eq k, Semigroup a) => [(k, a)] -> MonoidalMap k a
-fromAscList = fromAscListWith (<>)
+fromAscList :: forall k a. Eq k => [(k, a)] -> MonoidalMap k a
+fromAscList = coerce (M.fromAscList :: [(k, a)] -> M.Map k a)
 {-# INLINE fromAscList #-}
 
 fromAscListWith :: forall k a. Eq k => (a -> a -> a) -> [(k, a)] -> MonoidalMap k a
@@ -548,6 +573,22 @@ fromDistinctList :: forall k a. Ord k => [(k, a)] -> MonoidalMap k a
 fromDistinctList = coerce (M.fromList :: [(k, a)] -> M.Map k a)
 {-# INLINE fromDistinctList #-}
 
+fromDescList :: forall k a. Eq k => [(k, a)] -> MonoidalMap k a
+fromDescList = coerce (M.fromDescList :: [(k, a)] -> M.Map k a)
+{-# INLINE fromDescList #-}
+
+fromDescListWith :: forall k a. Eq k => (a -> a -> a) -> [(k, a)] -> MonoidalMap k a
+fromDescListWith = coerce (M.fromDescListWith :: (a -> a -> a) -> [(k, a)] -> M.Map k a)
+{-# INLINE fromDescListWith #-}
+
+fromDescListWithKey :: forall k a. Eq k => (k -> a -> a -> a) -> [(k, a)] -> MonoidalMap k a
+fromDescListWithKey = coerce (M.fromDescListWithKey :: (k -> a -> a -> a) -> [(k, a)] -> M.Map k a)
+{-# INLINE fromDescListWithKey #-}
+
+fromDistinctDescList :: forall k a. [(k, a)] -> MonoidalMap k a
+fromDistinctDescList = coerce (M.fromDistinctDescList :: [(k, a)] -> M.Map k a)
+{-# INLINE fromDistinctDescList #-}
+
 filter :: forall k a. (a -> Bool) -> MonoidalMap k a -> MonoidalMap k a
 filter = coerce (M.filter :: (a -> Bool) -> M.Map k a -> M.Map k a)
 {-# INLINE filter #-}
@@ -556,6 +597,14 @@ filterWithKey :: forall k a. (k -> a -> Bool) -> MonoidalMap k a -> MonoidalMap 
 filterWithKey = coerce (M.filterWithKey :: (k -> a -> Bool) -> M.Map k a -> M.Map k a)
 {-# INLINE filterWithKey #-}
 
+restrictKeys :: forall k a. Ord k => MonoidalMap k a -> Set k -> MonoidalMap k a
+restrictKeys = coerce (M.restrictKeys :: M.Map k a -> Set k -> M.Map k a)
+{-# INLINE restrictKeys #-}
+
+withoutKeys :: forall k a. Ord k => MonoidalMap k a -> Set k -> MonoidalMap k a
+withoutKeys = coerce (M.withoutKeys :: M.Map k a -> Set k -> M.Map k a)
+{-# INLINE withoutKeys #-}
+
 partition :: forall k a. (a -> Bool) -> MonoidalMap k a -> (MonoidalMap k a, MonoidalMap k a)
 partition = coerce (M.partition :: (a -> Bool) -> M.Map k a -> (M.Map k a, M.Map k a))
 {-# INLINE partition #-}
@@ -563,6 +612,18 @@ partition = coerce (M.partition :: (a -> Bool) -> M.Map k a -> (M.Map k a, M.Map
 partitionWithKey :: forall k a. (k -> a -> Bool) -> MonoidalMap k a -> (MonoidalMap k a, MonoidalMap k a)
 partitionWithKey = coerce (M.partitionWithKey :: (k -> a -> Bool) -> M.Map k a -> (M.Map k a, M.Map k a))
 {-# INLINE partitionWithKey #-}
+
+takeWhileAntitone :: forall k a. (k -> Bool) -> MonoidalMap k a -> MonoidalMap k a
+takeWhileAntitone = coerce (M.takeWhileAntitone :: (k -> Bool) -> M.Map k a -> M.Map k a)
+{-# INLINE takeWhileAntitone #-}
+
+dropWhileAntitone :: forall k a. (k -> Bool) -> MonoidalMap k a -> MonoidalMap k a
+dropWhileAntitone = coerce (M.dropWhileAntitone :: (k -> Bool) -> M.Map k a -> M.Map k a)
+{-# INLINE dropWhileAntitone #-}
+
+spanAntitone :: forall k a. (k -> Bool) -> MonoidalMap k a -> (MonoidalMap k a, MonoidalMap k a)
+spanAntitone = coerce (M.spanAntitone :: (k -> Bool) -> M.Map k a -> (M.Map k a, M.Map k a))
+{-# INLINE spanAntitone #-}
 
 mapMaybe :: forall k a b. (a -> Maybe b) -> MonoidalMap k a -> MonoidalMap k b
 mapMaybe = coerce (M.mapMaybe :: (a -> Maybe b) -> M.Map k a -> M.Map k b)
@@ -627,6 +688,26 @@ updateAt = coerce (M.updateAt :: (k -> a -> Maybe a) -> Int -> M.Map k a -> M.Ma
 deleteAt :: forall k a. Int -> MonoidalMap k a -> MonoidalMap k a
 deleteAt = coerce (M.deleteAt :: Int -> M.Map k a -> M.Map k a)
 {-# INLINE deleteAt #-}
+
+take :: forall k a. Int -> MonoidalMap k a -> MonoidalMap k a
+take = coerce (M.take :: Int -> M.Map k a -> M.Map k a)
+{-# INLINE take #-}
+
+drop :: forall k a. Int -> MonoidalMap k a -> MonoidalMap k a
+drop = coerce (M.drop :: Int -> M.Map k a -> M.Map k a)
+{-# INLINE drop #-}
+
+splitAt :: forall k a. Int -> MonoidalMap k a -> (MonoidalMap k a, MonoidalMap k a)
+splitAt = coerce (M.splitAt :: Int -> M.Map k a -> (M.Map k a, M.Map k a))
+{-# INLINE splitAt #-}
+
+lookupMin :: forall k a. MonoidalMap k a -> Maybe (k, a)
+lookupMin = coerce (M.lookupMin :: M.Map k a -> Maybe (k, a))
+{-# INLINE lookupMin #-}
+
+lookupMax :: forall k a. MonoidalMap k a -> Maybe (k, a)
+lookupMax = coerce (M.lookupMax :: M.Map k a -> Maybe (k, a))
+{-# INLINE lookupMax #-}
 
 findMin :: forall k a. MonoidalMap k a -> (k, a)
 findMin = coerce (M.findMin :: M.Map k a -> (k, a))