+++ Changelog.md
@@ -3,7 +3,7 @@
 # 0.6.1.0
 
   * Require containers 0.5.9 or greater.
-  * MonoidalMap: Add several functions added to Data.Map since 0.5 up to 0.5.9:
+  * MonoidalMap and MonoidalMap.Strict: Add several functions added to Data.Map since 0.5 up to 0.5.9:
     `take`, `drop`, `splitAt`, `lookupMin`, `lookupMax`, `restrictKeys`,
     `withoutKeys`, `fromDescList`,  `fromDescListWith`, `fromDescListWithKey`,
     `fromDistinctDescList`, `takeWhileAntitone`, `dropWhileAntitone`,